+++ src/views/Home.vue
@@ -12,12 +12,20 @@
             class="subtitle is-3"
           >Hindi(Inscript, Kruti Dev) and English Typing Tutor, Free for Windows</h2>
         </div>
-        <div class="container" v-show="isAvailableUpdate">
+        <div class="container">
           <br>
-          <div class="notification is-warning">
-            <strong style="cursor: pointer;" @click="openLink('https://github.com/ctechhindi/Speed-Typing-Tutor/releases')">
-              A new version available,
-            </strong> download the latest version and install in your pc.
+          <div v-show="isAvailableUpdate" class="notification is-warning">
+            <strong
+              style="cursor: pointer;"
+              @click="downloadUpdateFile()"
+            >A new version available,</strong> download the latest version and your new version will be automatically installed.
+          </div>
+          <div id="downloadContent" style="display: none;">
+            <p style="text-align: center; padding-bottom: 10px;">
+              <span style="color: #f93c3c; font-weight: 900;">(Please don't close application and Internet connection)</span>
+            </p>
+            <progress id="downloadProgess" class="progress is-success" value="0" max="100"></progress>
+            <p>Downloading...</p>
           </div>
         </div>
         <section class="section container is-fullhd">
@@ -100,16 +108,10 @@
             <a style="padding-right: 0.5em" @click="openLink($store.state.app.social.fb)">
               <b-icon size="is-large" type="is-primary" icon="facebook"></b-icon>
             </a>
-            <a
-              style="padding-right: 0.5em"
-              @click="openLink($store.state.app.social.instagram)"
-            >
+            <a style="padding-right: 0.5em" @click="openLink($store.state.app.social.instagram)">
               <b-icon size="is-large" type="is-primary" icon="instagram"></b-icon>
             </a>
-            <a
-              style="padding-right: 0.5em"
-              @click="openLink($store.state.app.social.github)"
-            >
+            <a style="padding-right: 0.5em" @click="openLink($store.state.app.social.github)">
               <b-icon size="is-large" type="is-primary" icon="github-circle"></b-icon>
             </a>
             <a
@@ -118,22 +120,18 @@
             >
               <b-icon size="is-large" type="is-primary" icon="stack-overflow"></b-icon>
             </a>
-            <a
-              style="padding-right: 0.5em"
-              @click="openLink($store.state.app.social.twitter)"
-            >
+            <a style="padding-right: 0.5em" @click="openLink($store.state.app.social.twitter)">
               <b-icon size="is-large" type="is-primary" icon="twitter"></b-icon>
             </a>
-            <a
-              style="padding-right: 0.5em"
-              @click="openLink($store.state.app.social.googlePlus)"
-            >
+            <a style="padding-right: 0.5em" @click="openLink($store.state.app.social.googlePlus)">
               <b-icon size="is-large" type="is-primary" icon="google-plus"></b-icon>
             </a>
           </p>
           <p>
             Copyrights ©
-            <a @click="openLink($store.state.app.copyrightURL)">{{ this.$store.state.app.copyright }}</a> 2018 | Made with
+            <a
+              @click="openLink($store.state.app.copyrightURL)"
+            >{{ this.$store.state.app.copyright }}</a> 2018 | Made with
             <b-icon icon="heart" type="is-danger"></b-icon>in India
           </p>
         </div>
@@ -143,16 +141,48 @@
 </template>
 
 <script>
-const APPP = require('../../package.json') // Load Package JSON File
-import axios from "axios";
+const APPP = require("../../package.json"); // Load Package JSON File
+
+import axios from "axios"
+const { remote, ipcRenderer } = require("electron")
+
+/**
+ * Application Update Download Progress
+ */
+ipcRenderer.on("download-progress", (event, d) => {
+  console.log(d);
+  if (d.receivedSize <= d.totalSize) {
+    document.getElementById('downloadProgess').setAttribute("max", d.totalSize);
+    document.getElementById('downloadProgess').setAttribute("value", d.receivedSize);
+    // document.getElementById('downloadProgess').style.display = 'none';
+  }
+});
+
+/**
+ * Application Update Download Complete
+ */
+ipcRenderer.on("download-complete", (event, msg) => {
+  console.log("download-complete");
+  // remote.app.relaunch({ args: process.argv.slice(1).concat(['--relaunch']) })
+  // remote.app.exit(0)
+});
+
+/**
+ * Application Update Download Failed
+ */
+ipcRenderer.on("download-failed", (event, msg) => {
+  alert("Application New Version Download is Failed.")
+});
+
 
 export default {
   name: "home",
   data() {
     return {
       // Check Available Update
-      isAvailableUpdate: false
-    }
+      isAvailableUpdate: true,
+      downloadURL: "https://raw.githubusercontent.com/ctechhindi/Speed-Typing-Tutor/master/docs/images/speed-typing-tutor.png",
+    };
   },
   methods: {
     /**
@@ -185,7 +215,7 @@ export default {
      * Check Application Update is Available
      */
     checkUpdate() {
-      var that = this
+      var that = this;
       axios
         .get(
           "https://raw.githubusercontent.com/ctechhindi/Speed-Typing-Tutor/master/package.json"
@@ -193,20 +223,31 @@ export default {
         .then(function(resp) {
           if (resp.data.version !== undefined) {
             if (that.compareVersions(resp.data.version, APPP.version)) {
-              that.isAvailableUpdate = true
+              that.isAvailableUpdate = true;
             }
           }
         })
         .catch(function(error) {
           console.log(error);
         });
-    }
+    },
+
+    /**
+     * Download Update File
+     */
+    downloadUpdateFile() {
+      ipcRenderer.send("update-download", {
+        url: this.downloadURL,
+      })
+      this.isAvailableUpdate = false
+      document.getElementById('downloadContent').style.display = 'block';
+    },
   },
   mounted() {
     this.$Progress.finish();
   },
   created() {
-    this.checkUpdate()
+    // this.checkUpdate()
   }
 };
 </script>